requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.3.0
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.5.0
python-dotenv==1.0.1
rich==13.7.1
openai>=1.12.0
